requirements.txt
aioredis==2.0.1
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.5.18.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.1.5
django-heroku==0.3.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
packaging==21.3
psycopg2==2.9.3
PyJWT==2.4.0
pymongo==3.12.3
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
sqlparse==0.2.4
typing_extensions==4.3.0
whitenoise==6.2.0
wrapt==1.14.1
